move LowPriorityStoreProvider to a dedicated file
